Unit tests EmailMessageBuilder
