+++ src/test/java/com/polifono/service/impl/SendEmailServiceTest.java
@@ -1,0 +1,14 @@
+package com.polifono.service.impl;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+public class SendEmailServiceTest {
+
+    @Test
+    public void givenPlayer_WhenSendEmailConfirmRegister_ThenSendEmail() {
+
+    }
+}
